Extract case handler screen out from app.js
